Changed the name of the packages so that they match java standards.
