--- src/Engine/Collision/CollisionDirection.java
@@ -1,14 +1,0 @@
-package Engine.Collision;
-
-import Engine.Engine;
-
-/**
- * Created by Maashes on 3/30/2016.
- */
-public enum CollisionDirection {
-    NONE,
-    FROMABOVE,
-    FROMBELOW,
-    FROMLEFT,
-    FROMRIGHT
-}
