--- src/Engine/Collision/CollisionType.java
@@ -1,14 +1,0 @@
-package Engine.Collision;
-
-/**
- * Created by Maashes on 4/12/2016.
- */
-public enum CollisionType
-{
-    ENEMYPROJECTILE,
-    ENEMY,
-    GROUND,
-    PLAYER,
-    PLAYERPROJECTILE,
-    ENVIRONMENT
-}
