--- src/Engine/EngineProvider.java
@@ -1,19 +1,0 @@
-package Engine;
-
-/**
- * Created by Maashes on 4/10/2016.
- */
-public class EngineProvider
-{
-    static Engine _engine;
-
-    public static Engine GetDefaultEngineInstance()
-    {
-        return _engine;
-    }
-
-    public static void SetDefaultEngineInstance(Engine e)
-    {
-        _engine = e;
-    }
-}
