--- src/GameObject/Animation/AnimationState.java
@@ -1,16 +1,0 @@
-package GameObject.Animation;
-
-import GameObject.GameObject;
-
-/**
- * Created by Maashes on 4/3/2016.
- */
-
-public enum AnimationState
-{
-    RUNNING,
-    JUMPING,
-    ATTACKING,
-    DAMAGED,
-    DEAD
-}
