--- src/GameObject/Enemy/EnemyType.java
@@ -1,12 +1,0 @@
-package GameObject.Enemy;
-
-/**
- * Created by Maashes on 4/10/2016.
- */
-public enum EnemyType
-{
-    RANGED,
-    MELEE,
-    BOSS,
-    BONUS
-}
