+++ src/com/shadowgame/engine/Collision/CollisionDirection.java
@@ -1,0 +1,12 @@
+package com.shadowgame.engine.Collision;
+
+/**
+ * Created by Maashes on 3/30/2016.
+ */
+public enum CollisionDirection {
+    NONE,
+    FROMABOVE,
+    FROMBELOW,
+    FROMLEFT,
+    FROMRIGHT
+}
