+++ src/com/shadowgame/engine/Collision/CollisionType.java
@@ -1,0 +1,14 @@
+package com.shadowgame.engine.Collision;
+
+/**
+ * Created by Maashes on 4/12/2016.
+ */
+public enum CollisionType
+{
+    ENEMYPROJECTILE,
+    ENEMY,
+    GROUND,
+    PLAYER,
+    PLAYERPROJECTILE,
+    ENVIRONMENT
+}
