+++ src/com/shadowgame/engine/EngineProvider.java
@@ -1,0 +1,19 @@
+package com.shadowgame.engine;
+
+/**
+ * Created by Maashes on 4/10/2016.
+ */
+public class EngineProvider
+{
+    static Engine _engine;
+
+    public static Engine GetDefaultEngineInstance()
+    {
+        return _engine;
+    }
+
+    public static void SetDefaultEngineInstance(Engine e)
+    {
+        _engine = e;
+    }
+}
