+++ src/com/shadowgame/gameobject/animation/AnimationState.java
@@ -1,0 +1,14 @@
+package com.shadowgame.gameobject.animation;
+
+/**
+ * Created by Maashes on 4/3/2016.
+ */
+
+public enum AnimationState
+{
+    RUNNING,
+    JUMPING,
+    ATTACKING,
+    DAMAGED,
+    DEAD
+}
