+++ src/com/shadowgame/gameobject/enemy/EnemyType.java
@@ -1,0 +1,12 @@
+package com.shadowgame.gameobject.enemy;
+
+/**
+ * Created by Maashes on 4/10/2016.
+ */
+public enum EnemyType
+{
+    RANGED,
+    MELEE,
+    BOSS,
+    BONUS
+}
